Fix QuicNetworkTransaction*Tests to set QuicAlternativeService with current supported versions in the test suite.
Currently we ignore QUIC versions when GetAlternativeService. However,
in order to use the AlternativeService properly, a correct list of
versions in AlternativeServiceInfo should be set to HttpServerProperties
which contains the QUIC version that the running test is supporting.

This is in preparation to fix the tests that would blow up once we take
QUIC versions into account when selecting AlternativeService.

BUG=689762

Review-Url: https://codereview.chromium.org/2966563003
Cr-Original-Commit-Position: refs/heads/master@{#483575}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 86838d5ee0c79c95644607bc0c29a7dcc54482b3