Fix MockHostResolver calling system resolver with invalid DNS values.
HostResolverImpl used to pass all non-numeric host names it received,
including invalid ones, to the system HostResolver. That was changed
in https://chromium-review.googlesource.com/569298 (Since the bad
addresses could be sent over the wire with the async resolver, and
it generally seemed like a bad idea).  And all was right with the world.

Then a DCHECK was added to SystemHostResolverCall for invalid hostnames,
since it should no longer be called with them in production. However,
it turns out the MockHostResolver calls that method directly, bypassing
HostResolverImpl, and does so with IPv6 host names (Or whatever other
stuff is passed to it). In particular, the socks_client_socket_fuzzer
does this by passing in "::1".  This is a legal use of the
MockHostResolver API, not a bug in the fuzzer.  So this CL modifies
MockHostResolver not to do that.

Bug: 747243
Change-Id: Ib4be76f1d03f33bc65a493ac73aae6407ca3383d
Reviewed-on: https://chromium-review.googlesource.com/581788
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488800}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ac004d5921840187883c331e161fb3bc5ae880f