android: Test that preconnects can be issued from the Java side.
Preconnects can be initiated from WarmupManager. This adds the testing
infrastructure and a basic test checking that these preconnects are
actually issued. This is required to test the redirect preconnection
handling for Custom Tabs.
Forthcoming CLs will add tests to check that the sockets are re-used,
and end-to-end tests of redirect preconnection.

In terms of implementation:
Native:
  - Sets an EmbeddedTestServerConnectionListener on the native side.
  - Have this listener relay connection data to the Java side.

In Java:
  - Add a callback AIDL interface to relay callbacks from the server to
    the test APK.
  - Have the client-side callback post a task to the UI thread, as AIDL
    backed callbacks are delivered to the binder thread pool.

Bug: 739165
Change-Id: Ifa461aa614354eaa5cb5d7b1aec5ba84b8d1b8f7
Reviewed-on: https://chromium-review.googlesource.com/583607
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Misha Efimov <mef@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c5197808ef235ba93e0376b6d8288b12a7e26be