Emit a warning when NS Server Gated Crypto is used in place of server auth in a certificate chain.
Netscape Server Gated Crypto (2.16.840.1.113730.4.1) is a deprecated
mechanism that is still in use by some unexpired (intermediate)
certificates. It is not part of the RFC 5280 profile for internet PKI.

This change adds a warning for the chain(s) in the certificate that are relying on nsSGC in place of server auth.

Bug: 733403
Change-Id: I2994f7f5e2981eecf24bb9839e266d4292cad7f8
Reviewed-on: https://chromium-review.googlesource.com/593207
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Eric Roman <eroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491141}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c8c2d6a74bae4340f86e3c8889aeb04fdc9b6bfc