Log console message for legacy Symantec roots
This CL implements the console message described in https://security.googleblog.com/2017/09/chromes-plan-to-distrust-symantec.html
for main resources. To do so, it:
1.) Relocates the existing list of Symantec roots (currently only used
for CT policies) and exposes it as
net::IsLegacySymantecCert.
2.) Stores public key hashes from the net::SSLInfo on navigation
entries and checks them against IsLegacySymantecCert when a
navigation commits. When a match is found, log a message to
the DevTools console.

Note that this message will not fire for resources served from the
disk cache, because public_key_hashes are not written to the disk
cache.

Bug: 763984
Change-Id: I013a1226c63a0192569192839f30bfbb22afffaa
Reviewed-on: https://chromium-review.googlesource.com/659318
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502472}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d29cdae2baf037fe2fd3e761c69ef498c8100ae0