Fix an issue that EmbeddedTestServer may not send headers that can be parsed correctly
HttpUtils is looking for "\n\n" and "\n\r\n" as the end of the header.
So if the headers are already ending with "\n\n", we shouldn't append "\r\n" to it.
Otherwise, the appended "\r\n" will be treated as the body

BUG=715630

Change-Id: Ia087f3ec7f7644c088635f375782841f2c6a32d0
Reviewed-on: https://chromium-review.googlesource.com/693323
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32c60fdc1d049e8b78bc59c0770eda996d976944