SimpleCache: Permit optimistic create while doom is pending.
This is desirable if we use in memory hints to avoid opening files in CantConditionalize
case --- in that case we still want a cache entry, so the simplest way of doing that is
to Doom the old one and Create a new one, which before this change would block until
Doom completes. This lets CreateEntry return an Entry object optimistically, as long as 
the create is the first operation after Doom. 

This works by creating the entry in STATE_IO_PENDING state, and transitioning away from
that and actually starting creation I/O via an entries_pending_doom_ callback.

(This may be also be me trying to be too clever, I am worried this may be a bit too cute)

Bug: 729679
Change-Id: I210f7ae5686a1a04e0930ea5c13e67c81131bd1a
Reviewed-on: https://chromium-review.googlesource.com/684855
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Julia Tuttle <juliatuttle@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#512151}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6d2d88636508af23b09cec87bbbd57af4778407