Disable URLRequestFileDirTest tests that fail on Android.
Various RootDirectory test cases exist to make sure browsing file:///
works. However, this action is not allowed on Android, so it is a slight
mystery how the tests manage to pass on earlier version of Android. The
tests fail under newer version of Android.

Given this situation, just disable the tests on Android altogether, as
it is pointless to test a disallowed action. Also combine the test cases
which are very similar into a single test case with a for-loop.

BUG=787149

Change-Id: Icec5ed676b245519ae98e8914b397122d5c7bbd5
Reviewed-on: https://chromium-review.googlesource.com/780921
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 284aae98c95fd4a60e0d1b6769b65fd21f1de29f