Reporting API: Remove old endpoints when processing header.
Origins shouldn't need to remember and explicitly remove every
deprecated endpoint they've used in the past.

This change alters the header parsing behavior so any endpoint that was
configured for an origin but is not included in a new header from that
origin will be removed.

Change-Id: I87590e00340684ad3deaefe5ff2a2939d904a983
Reviewed-on: https://chromium-review.googlesource.com/797737
Commit-Queue: Julia Tuttle <juliatuttle@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#521355}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 443a0a6831d02638ef248b14ce841760e2df0ecf